fix: coderabbit 오류 수정, point관련 알림 라우팅 경로 수정
