+++ src/pages/coffee-chat/components/ChatList.tsx
@@ -1,13 +1,6 @@
 import type { ChatRoomListItem } from "../../../types/coffee-chat/coffeeChatTypes";
 import { formatDate } from "../../../utils/formatDate";
-
-const formatStudentLabel = (studentId?: string) => {
-    const normalized = studentId?.trim() ?? "";
-    if (!normalized) return "";
-    return !isNaN(Number(normalized)) && normalized.length >= 2
-        ? `${normalized.slice(2, 4)}학번`
-        : normalized;
-};
+import { formatStudentLabel } from "../../../utils/formatStudent";
 
 interface ChatListProps {
     chatRoom: ChatRoomListItem;
