+++ src/pages/home/notificationRouting.ts
@@ -1,0 +1,45 @@
+import type { NotificationItem } from './notificationData';
+
+export const buildCommunityPostLink = (postId?: number, commentId?: number) => {
+  if (!postId) return null;
+  if (!commentId) return `/community/post/${postId}`;
+
+  const searchParams = new URLSearchParams({
+    commentId: String(commentId),
+  });
+
+  return `/community/post/${postId}?${searchParams.toString()}`;
+};
+
+export const resolveNotificationDestination = (notification: NotificationItem) => {
+  if (notification.link) return notification.link;
+
+  switch (notification.type) {
+    case 'coffeeChatRequest':
+    case 'teamApplicationReceived':
+      return notification.requestId
+        ? `/chat/requests/${notification.requestId}`
+        : '/chat/requests';
+    case 'coffeeChatAccepted':
+    case 'chatMessageReceived':
+      return '/chat';
+    case 'teamRecruitAccepted':
+      return notification.requestId
+        ? `/chat/requests/${notification.requestId}`
+        : '/chat/requests';
+    case 'followingPosted':
+      return notification.postId
+        ? buildCommunityPostLink(notification.postId, notification.commentId)
+        : '/community';
+    case 'commentAccepted':
+    case 'reply':
+    case 'comment':
+      return buildCommunityPostLink(notification.postId, notification.commentId);
+    case 'pointUse':
+    case 'pointEarn':
+      return '/shop';
+    case 'default':
+    default:
+      return null;
+  }
+};
